// client/src/models/items/potions/potion.types.ts
export enum PotionType {
  COAL_POTION = 'coal potion',
  BRONZE_POTION = 'bronze potion',
  IRON_POTION = 'iron potion',
  GOLD_POTION = 'gold potion',
  TURQUOISE_POTION = 'turquoise potion',
  ZAPHIRE_POTION = 'zaphire potion',
  EMERALD_POTION = 'emerald potion',
  RUBI_POTION = 'rubi potion',

  COAL_POWDER = 'coal powder',
  BRONZE_POWDER = 'bronze powder',
  IRON_POWDER = 'iron powder',
  GOLD_POWDER = 'gold powder',
  TURQUOISE_POWDER = 'turquoise powder',
  ZAPHIRE_POWDER = 'zaphire powder',
  EMERALD_POWDER = 'emerald powder',
  RUBI_POWDER = 'rubi powder',

  COAL_SPELL = 'coal spell',
  BRONZE_SPELL = 'bronze spell',
  IRON_SPELL = 'iron spell',
  GOLD_SPELL = 'gold spell',
  TURQUOISE_SPELL = 'turquoise spell',
  ZAPHIRE_SPELL = 'zaphire spell',
  EMERALD_SPELL = 'emerald spell',
  RUBI_SPELL = 'rubi spell',

  COAL_TEAR = 'coal tear',
  BRONZE_TEAR = 'bronze tear',
  IRON_TEAR = 'iron tear',
  GOLD_TEAR = 'gold tear',
  TURQUOISE_TEAR = 'turquoise tear',
  ZAPHIRE_TEAR = 'zaphire tear',
  EMERALD_TEAR = 'emerald tear',
  RUBI_TEAR = 'rubi tear',

  COAL_RAGE = 'coal rage',
  BRONZE_RAGE = 'bronze rage',
  IRON_RAGE = 'iron rage',
  GOLD_RAGE = 'gold rage',
  TURQUOISE_RAGE = 'turquoise rage',
  ZAPHIRE_RAGE = 'zaphire rage',
  EMERALD_RAGE = 'emerald rage',
  RUBI_RAGE = 'rubi rage',

  COAL_WISH = 'coal wish',
  BRONZE_WISH = 'bronze wish',
  IRON_WISH = 'iron wish',
  GOLD_WISH = 'gold wish',
  TURQUOISE_WISH = 'turquoise wish',
  ZAPHIRE_WISH = 'zaphire wish',
  EMERALD_WISH = 'emerald wish',
  RUBI_WISH = 'rubi wish',
}